Migrate to AGP 9 and new project structure, and upgrade dependencies
