+++ build.gradle.kts
@@ -2,10 +2,10 @@ plugins {
     // this is necessary to avoid the plugins to be loaded multiple times
     // in each subproject's classloader
     alias(libs.plugins.androidApplication) apply false
-    alias(libs.plugins.androidLibrary) apply false
+    alias(libs.plugins.androidMultiplatformLibrary) apply false
     alias(libs.plugins.composeMultiplatform) apply false
     alias(libs.plugins.composeCompiler) apply false
     alias(libs.plugins.kotlinMultiplatform) apply false
-    id("com.google.devtools.ksp").version("2.2.20-2.0.3").apply(false)
-    id("com.rickclephas.kmp.nativecoroutines").version("1.0.0-ALPHA-47").apply(false)
+    alias(libs.plugins.kotlinSerialization) apply false
+    alias(libs.plugins.kmpNativeCoroutines) apply false
 }
