--- composeApp/src/androidUnitTest/kotlin/com/jetbrains/greeting/greetingkmp/ComposeAppAndroidUnitTest.kt
@@ -1,12 +1,0 @@
-package com.jetbrains.greeting.greetingkmp
-
-import kotlin.test.Test
-import kotlin.test.assertEquals
-
-class ComposeAppAndroidUnitTest {
-
-    @Test
-    fun example() {
-        assertEquals(3, 1 + 2)
-    }
-}
