+++ iosApp/iosApp.xcodeproj/project.xcworkspace/xcshareddata/swiftpm/Package.resolved
@@ -6,8 +6,8 @@
       "kind" : "remoteSourceControl",
       "location" : "https://github.com/rickclephas/KMP-NativeCoroutines.git",
       "state" : {
-        "revision" : "131f600110c4ba215585cb03ec5d7b6ad1c1fde2",
-        "version" : "1.0.0-ALPHA-45"
+        "revision" : "ba036458958d28b24c623f61700b1cb9e9b4e82c",
+        "version" : "1.0.1"
       }
     },
     {
