+++ settings.gradle.kts
@@ -28,5 +28,6 @@ dependencyResolutionManagement {
     }
 }
 
-include(":composeApp")
-include(":shared")
+include(":androidApp")
+include(":sharedLogic")
+include(":sharedUI")
