--- shared/src/commonTest/kotlin/com/jetbrains/greeting/greetingkmp/SharedCommonTest.kt
@@ -1,12 +1,0 @@
-package com.jetbrains.greeting.greetingkmp
-
-import kotlin.test.Test
-import kotlin.test.assertEquals
-
-class SharedCommonTest {
-
-    @Test
-    fun example() {
-        assertEquals(3, 1 + 2)
-    }
-}
