+++ sharedLogic/src/androidMain/kotlin/com/jetbrains/greeting/greetingkmp/Platform.android.kt
