+++ sharedLogic/src/commonMain/kotlin/com/jetbrains/greeting/greetingkmp/NewYear.kt
