+++ sharedLogic/src/commonMain/kotlin/com/jetbrains/greeting/greetingkmp/Platform.kt
