+++ sharedLogic/src/commonMain/kotlin/com/jetbrains/greeting/greetingkmp/RocketComponent.kt
