+++ sharedLogic/src/commonMain/kotlin/com/jetbrains/greeting/greetingkmp/RocketLaunch.kt
