+++ sharedLogic/src/commonTest/kotlin/com/jetbrains/greeting/greetingkmp/SharedLogicCommonTest.kt
@@ -1,0 +1,12 @@
+package com.jetbrains.greeting.greetingkmp
+
+import kotlin.test.Test
+import kotlin.test.assertEquals
+
+class SharedLogicCommonTest {
+
+    @Test
+    fun example() {
+        assertEquals(3, 1 + 2)
+    }
+}
