+++ sharedLogic/src/iosMain/kotlin/com/jetbrains/greeting/greetingkmp/Platform.ios.kt
