+++ sharedUI/src/commonMain/kotlin/com/jetbrains/greeting/greetingkmp/MainViewModel.kt
